Move and clean up test
