+++ tests/test_thirdreality.py
@@ -1,13 +1,34 @@
 """Tests for Third Reality quirks."""
 
+from unittest import mock
+
 import pytest
 from zigpy.zcl.clusters.security import IasZone
 
 from tests.common import ClusterListener
 import zhaquirks
+from zhaquirks.thirdreality.button import MultistateInputCluster
 import zhaquirks.thirdreality.night_light
 
 zhaquirks.setup()
+
+
+class MockListener:
+    """Simulate listener class for capturing ZHA events."""
+
+    def __init__(self):
+        """Initialize listener with empty event list."""
+        self.zha_send_events = []
+
+    def zha_send_event(self, action, event_args):
+        """Record ZHA events.
+
+        Args:
+            action (str): The type of action for the event.
+            event_args (dict): Relevant parameters of the event.
+
+        """
+        self.zha_send_events.append((action, event_args))
 
 
 @pytest.mark.parametrize("quirk", (zhaquirks.thirdreality.night_light.Nightlight,))
@@ -36,3 +57,38 @@ async def test_third_reality_nightlight(zigpy_device_from_quirk, quirk):
     assert len(ias_zone_listener.attribute_updates) == 2
     assert ias_zone_listener.attribute_updates[1][0] == ias_zone_status_id
     assert ias_zone_listener.attribute_updates[1][1] == 0
+
+
+@pytest.mark.parametrize(
+    ("attr_value", "expected_action"),
+    [
+        (1, "single"),  # 1 corresponds to single click
+        (2, "double"),  # 2 corresponds to double click
+        (0, "hold"),  # 0 corresponds to hold
+        (255, "release"),  # 255 corresponds to release
+    ],
+)
+@pytest.mark.parametrize(
+    ("manufacturer", "model"),
+    [("Third Reality, Inc", "3RSB22BZ")],
+)
+async def test_third_reality_button_v2(
+    zigpy_device_from_v2_quirk, manufacturer, model, attr_value, expected_action
+):
+    """Test Third Reality button event conversion and triggering functionality."""
+    device = zigpy_device_from_v2_quirk(manufacturer, model)
+    multistate_cluster = device.endpoints[1].in_clusters[
+        MultistateInputCluster.cluster_id
+    ]
+
+    # Create mock listener and register it with the cluster
+    listener = mock.MagicMock()
+    multistate_cluster.add_listener(listener)
+
+    multistate_cluster.update_attribute(
+        0x0055, attr_value
+    )  # 1 corresponds to single click
+    assert listener.zha_send_event.call_count == 1
+    assert listener.zha_send_event.call_args_list[0] == mock.call(
+        expected_action, {"value": attr_value}
+    )
